+++ Changelog.md
@@ -2,7 +2,9 @@
 
 ## master (unreleased)
 
-Nothing here yet.
+- Removed duplicate Proclamation Day for Latvia, by @Daglina (#686).
+- New calendar: Added Kazakhstan calendar by @atj01 (#688).
+- New calendar: Added Georgia (country) calendar by @atj01 (#687).
 
 ## v16.2.0 (2021-12-10)